(MacOS) Fix Windowing bug in Logic
In ~WindowImplCocoa(), SFML was attempting to make the next ordered window key and front. This was causing a bug in Logic where a window that was supposed to be destroyed became a zombie and wouldn't go away. In general, it seems like we should allow the OS or calling function to decide what to do next, or which window to focus after closing a window.